Moved TaskThread implementation to Application class
